combined hooks
